+++ CHANGELOG.md
@@ -46,7 +46,9 @@
       on that order will now break.
       This also changes the output of
       `AddToHueAndSaturation.get_parameters()`. #319
+* Added `AddToHue`, a shortcut for `AddToHueAndSaturation(value_hue=...)`. #319
 * Refactored `augmenters/weather.py` (general code and docstring cleanup). #336
+
 
 ## Fixes